+++ src/main/kotlin/io/verik/intellij/inspection/common/VerikInspectionProvider.kt
@@ -18,6 +18,7 @@ package io.verik.intellij.inspection.common
 
 import com.intellij.codeInspection.InspectionToolProvider
 import com.intellij.codeInspection.LocalInspectionTool
+import io.verik.intellij.inspection.inspection.FileAnnotationInspection
 import io.verik.intellij.inspection.inspection.ModuleNotClass
 import io.verik.intellij.inspection.inspection.ModuleNotObject
 import io.verik.intellij.inspection.inspection.TopNotModuleInspection
@@ -28,6 +29,7 @@ class VerikInspectionProvider : InspectionToolProvider {
 
     override fun getInspectionClasses(): Array<Class<out LocalInspectionTool>> {
         return arrayOf(
+            FileAnnotationInspection::class.java,
             UnsupportedElementInspection::class.java,
             UnsupportedModifierInspection::class.java,
             TopNotModuleInspection::class.java,